Add gRPC token based authentication test and move tests to another file
