+++ gateway/cert_test.go
@@ -2,12 +2,10 @@ package gateway
 
 import (
 	"bytes"
-	"context"
 	"crypto/rand"
 	"crypto/rsa"
 	"crypto/tls"
 	"crypto/x509"
-	"encoding/base64"
 	"encoding/pem"
 	"fmt"
 	"io/ioutil"
@@ -17,16 +15,8 @@ import (
 	"net/http/httptest"
 	"os"
 	"path/filepath"
-	"strings"
 	"testing"
 	"time"
-
-	"google.golang.org/grpc"
-	pb "google.golang.org/grpc/examples/helloworld/helloworld"
-
-	"google.golang.org/grpc/credentials"
-
-	"golang.org/x/net/http2"
 
 	"github.com/TykTechnologies/tyk/apidef"
 	"github.com/TykTechnologies/tyk/certs"
@@ -657,326 +647,4 @@ func TestCipherSuites(t *testing.T) {
 
 		ts.Run(t, test.TestCase{Client: client, Path: "/", ErrorMatch: "tls: handshake failure"})
 	})
-}
-
-func TestHTTP2(t *testing.T) {
-	expected := "HTTP/2.0"
-
-	// Certificates
-	_, _, _, clientCert := genCertificate(&x509.Certificate{})
-	serverCertPem, _, combinedPEM, _ := genServerCertificate()
-	certID, _ := CertificateManager.Add(combinedPEM, "")
-	defer CertificateManager.Delete(certID)
-
-	// Upstream server supporting HTTP/2
-	upstream := httptest.NewUnstartedServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		actual := r.Proto
-		if expected != actual {
-			t.Fatalf("Tyk-Upstream connection protocol is expected %s, actual %s", expected, actual)
-		}
-
-		fmt.Fprintln(w, "Hello, I am an HTTP/2 Server")
-
-	}))
-	upstream.TLS = new(tls.Config)
-	upstream.TLS.NextProtos = []string{"h2"}
-	upstream.StartTLS()
-	defer upstream.Close()
-
-	// Tyk
-	globalConf := config.Global()
-	globalConf.ProxySSLInsecureSkipVerify = true
-	globalConf.ProxyEnableHttp2 = true
-	globalConf.HttpServerOptions.EnableHttp2 = true
-	globalConf.HttpServerOptions.SSLCertificates = []string{certID}
-	globalConf.HttpServerOptions.UseSSL = true
-	config.SetGlobal(globalConf)
-	defer ResetTestConfig()
-
-	ts := StartTest()
-	defer ts.Close()
-
-	BuildAndLoadAPI(func(spec *APISpec) {
-		spec.Proxy.ListenPath = "/"
-		spec.UseKeylessAccess = true
-		spec.Proxy.TargetURL = upstream.URL
-	})
-
-	// HTTP/2 client
-	http2Client := getTLSClient(&clientCert, serverCertPem)
-	http2.ConfigureTransport(http2Client.Transport.(*http.Transport))
-
-	ts.Run(t, test.TestCase{Client: http2Client, Path: "", Code: 200, Proto: "HTTP/2.0", BodyMatch: "Hello, I am an HTTP/2 Server"})
-}
-
-// server is used to implement helloworld.GreeterServer.
-type server struct{}
-
-// SayHello implements helloworld.GreeterServer
-func (s *server) SayHello(ctx context.Context, in *pb.HelloRequest) (*pb.HelloReply, error) {
-	log.Printf("Received: %v", in.Name)
-	return &pb.HelloReply{Message: "Hello " + in.Name}, nil
-}
-
-func TestGRPC_TLS(t *testing.T) {
-
-	_, _, combinedPEM, _ := genServerCertificate()
-	certID, _ := CertificateManager.Add(combinedPEM, "")
-	defer CertificateManager.Delete(certID)
-
-	// gRPC server
-	s := startGRPCServer(t, nil)
-	defer s.GracefulStop()
-
-	// Tyk
-	globalConf := config.Global()
-	globalConf.ProxySSLInsecureSkipVerify = true
-	globalConf.ProxyEnableHttp2 = true
-	globalConf.HttpServerOptions.EnableHttp2 = true
-	globalConf.HttpServerOptions.SSLCertificates = []string{certID}
-	globalConf.HttpServerOptions.UseSSL = true
-	config.SetGlobal(globalConf)
-	defer ResetTestConfig()
-
-	ts := StartTest()
-	defer ts.Close()
-
-	BuildAndLoadAPI(func(spec *APISpec) {
-		spec.Proxy.ListenPath = "/"
-		spec.UseKeylessAccess = true
-		spec.Proxy.TargetURL = "https://localhost:50051"
-	})
-
-	address := strings.TrimPrefix(ts.URL, "https://")
-	name := "Furkan"
-
-	// gRPC client
-	r := sayHelloWithGRPCClient(t, nil, nil, false, address, name)
-
-	// Test result
-	expected := "Hello " + name
-	actual := r.Message
-
-	if expected != actual {
-		t.Fatalf("Expected %s, actual %s", expected, actual)
-	}
-}
-
-func TestGRPC_MutualTLS(t *testing.T) {
-	// Mutual Authentication for both downstream-tyk and tyk-upstream
-
-	_, _, combinedClientPEM, clientCert := genCertificate(&x509.Certificate{})
-	clientCert.Leaf, _ = x509.ParseCertificate(clientCert.Certificate[0])
-	serverCertPem, _, combinedPEM, _ := genServerCertificate()
-
-	certID, _ := CertificateManager.Add(combinedPEM, "") // For tyk to know downstream
-	defer CertificateManager.Delete(certID)
-
-	clientCertID, _ := CertificateManager.Add(combinedClientPEM, "") // For upstream to know tyk
-	defer CertificateManager.Delete(clientCertID)
-
-	// Protected gRPC server
-	s := startGRPCServer(t, clientCert.Leaf)
-	defer s.GracefulStop()
-
-	// Tyk
-	globalConf := config.Global()
-	globalConf.ProxySSLInsecureSkipVerify = true
-	globalConf.ProxyEnableHttp2 = true
-	globalConf.HttpServerOptions.EnableHttp2 = true
-	globalConf.HttpServerOptions.SSLCertificates = []string{certID}
-	globalConf.HttpServerOptions.UseSSL = true
-	config.SetGlobal(globalConf)
-	defer resetTestConfig()
-
-	ts := newTykTestServer()
-	defer ts.Close()
-
-	buildAndLoadAPI(func(spec *APISpec) {
-		spec.Proxy.ListenPath = "/"
-		spec.UseKeylessAccess = true
-		spec.UpstreamCertificates = map[string]string{
-			"*": clientCertID,
-		}
-		spec.Proxy.TargetURL = "https://localhost:50051"
-	})
-
-	address := strings.TrimPrefix(ts.URL, "https://")
-	name := "Furkan"
-
-	// gRPC client
-	r := sayHelloWithGRPCClient(t, &clientCert, serverCertPem, false, address, name)
-
-	// Test result
-	expected := "Hello " + name
-	actual := r.Message
-
-	if expected != actual {
-		t.Fatalf("Expected %s, actual %s", expected, actual)
-	}
-}
-
-func startGRPCServer(t *testing.T, clientCert *x509.Certificate) *grpc.Server {
-	// Server
-	lis, err := net.Listen("tcp", ":50051")
-	if err != nil {
-		t.Fatalf("failed to listen: %v", err)
-	}
-
-	cert, key, _, _ := genCertificate(&x509.Certificate{})
-	certificate, _ := tls.X509KeyPair(cert, key)
-
-	pool := x509.NewCertPool()
-
-	tlsConfig := &tls.Config{}
-	if clientCert != nil {
-		tlsConfig = &tls.Config{
-			ClientAuth:         tls.RequireAndVerifyClientCert,
-			ClientCAs:          pool,
-			InsecureSkipVerify: true,
-			Certificates:       []tls.Certificate{certificate},
-		}
-		pool.AddCert(clientCert)
-	} else {
-		tlsConfig = &tls.Config{
-			InsecureSkipVerify: true,
-			Certificates:       []tls.Certificate{certificate},
-		}
-	}
-
-	creds := credentials.NewTLS(tlsConfig)
-
-	if err != nil {
-		t.Fatalf("failed to listen: %v", err)
-	}
-
-	s := grpc.NewServer(grpc.Creds(creds))
-
-	pb.RegisterGreeterServer(s, &server{})
-
-	go func() {
-		err := s.Serve(lis)
-		if err != nil {
-			t.Fatalf("failed to serve: %v", err)
-		}
-	}()
-
-	return s
-}
-
-func sayHelloWithGRPCClient(t *testing.T, cert *tls.Certificate, caCert []byte, basicAuth bool, address string, name string) *pb.HelloReply {
-	// gRPC client
-	tlsConfig := &tls.Config{}
-
-	if cert != nil {
-		tlsConfig.Certificates = []tls.Certificate{*cert}
-	}
-
-	if len(caCert) > 0 {
-		caCertPool := x509.NewCertPool()
-		caCertPool.AppendCertsFromPEM(caCert)
-		tlsConfig.RootCAs = caCertPool
-		tlsConfig.BuildNameToCertificate()
-	} else {
-		tlsConfig.InsecureSkipVerify = true
-	}
-
-	creds := credentials.NewTLS(tlsConfig)
-
-	opts := []grpc.DialOption{grpc.WithTransportCredentials(creds)}
-
-	if basicAuth {
-		opts = append(opts, grpc.WithPerRPCCredentials(&loginCreds{
-			Username: "user",
-			Password: "password",
-		}))
-	}
-
-	conn, err := grpc.Dial(address, opts...)
-	if err != nil {
-		t.Fatalf("did not connect: %v", err)
-	}
-	defer conn.Close()
-	c := pb.NewGreeterClient(conn)
-
-	ctx, cancel := context.WithTimeout(context.Background(), time.Second)
-	defer cancel()
-	r, err := c.SayHello(ctx, &pb.HelloRequest{Name: name})
-	if err != nil {
-		t.Fatalf("could not greet: %v", err)
-	}
-
-	return r
-}
-
-type loginCreds struct {
-	Username, Password string
-}
-
-func (c *loginCreds) GetRequestMetadata(context.Context, ...string) (map[string]string, error) {
-	auth := c.Username + ":" + c.Password
-	enc := base64.StdEncoding.EncodeToString([]byte(auth))
-	return map[string]string{
-		"Authorization": "Basic " + enc,
-	}, nil
-}
-
-func (c *loginCreds) RequireTransportSecurity() bool {
-	return true
-}
-
-func TestGRPC_BasicAuthentication(t *testing.T) {
-	_, _, combinedPEM, _ := genServerCertificate()
-	certID, _ := CertificateManager.Add(combinedPEM, "")
-	defer CertificateManager.Delete(certID)
-
-	// gRPC server
-	s := startGRPCServer(t, nil)
-	defer s.GracefulStop()
-
-	// Tyk
-	globalConf := config.Global()
-	globalConf.ProxySSLInsecureSkipVerify = true
-	globalConf.ProxyEnableHttp2 = true
-	globalConf.HttpServerOptions.EnableHttp2 = true
-	globalConf.HttpServerOptions.SSLCertificates = []string{certID}
-	globalConf.HttpServerOptions.UseSSL = true
-	config.SetGlobal(globalConf)
-	defer resetTestConfig()
-
-	ts := newTykTestServer()
-	defer ts.Close()
-
-	session := createStandardSession()
-	session.BasicAuthData.Password = "password"
-	session.AccessRights = map[string]user.AccessDefinition{"test": {APIID: "test", Versions: []string{"v1"}}}
-	session.OrgID = "default"
-
-	buildAndLoadAPI(func(spec *APISpec) {
-		spec.UseBasicAuth = true
-		spec.Proxy.ListenPath = "/"
-		spec.UseKeylessAccess = false
-		spec.Proxy.TargetURL = "https://localhost:50051"
-		spec.OrgID = "default"
-	})
-
-	address := strings.TrimPrefix(ts.URL, "https://")
-	name := "Furkan"
-	client := getTLSClient(nil, nil)
-
-	// To create key
-	ts.Run(t, []test.TestCase{
-		{Method: "POST", Path: "/tyk/keys/defaultuser", Data: session, AdminAuth: true, Code: 200, Client: client},
-	}...)
-
-	// gRPC client
-	r := sayHelloWithGRPCClient(t, nil, nil, true, address, name)
-
-	// Test result
-	expected := "Hello " + name
-	actual := r.Message
-
-	if expected != actual {
-		t.Fatalf("Expected %s, actual %s", expected, actual)
-	}
 }
