+++ trace/jaeger/jaeger.go
@@ -3,7 +3,7 @@ package jaeger
 import (
 	"io"
 
-	"github.com/opentracing/opentracing-go"
+	opentracing "github.com/opentracing/opentracing-go"
 	"github.com/uber/jaeger-client-go/config"
 )
